+++ puffin_http/examples/server.rs
@@ -7,16 +7,36 @@ fn main() {
     let server_addr = format!("0.0.0.0:{}", puffin_http::DEFAULT_PORT);
     eprintln!("Serving demo profile data on {}", server_addr);
 
-    let puffin_server = puffin_http::Server::new(&server_addr).unwrap();
+    let _puffin_server = puffin_http::Server::new(&server_addr).unwrap();
 
     puffin::set_scopes_on(true);
+
+    let mut frame_counter = 0;
 
     loop {
         puffin::profile_scope!("main_loop");
         puffin::GlobalProfiler::lock().new_frame();
-        puffin_server.update();
 
-        sleep_ms(16);
+        // Give us something to inspect:
+
+        std::thread::Builder::new()
+            .name("Other thread".to_owned())
+            .spawn(|| {
+                sleep_ms(5);
+            })
+            .unwrap();
+
+        sleep_ms(14);
+        if frame_counter % 7 == 0 {
+            puffin::profile_scope!("Spike");
+            std::thread::sleep(std::time::Duration::from_millis(10))
+        }
+
+        for _ in 0..1000 {
+            puffin::profile_scope!("very thin");
+        }
+
+        frame_counter += 1;
     }
 }